refactor(script): split CSV buildins into buildin_csv.c
Issue #61 eighth step: move CSV helpers out of script.c; share
csvinit/final and getarraysize via script_internal.h; replace
csvreadarray/csvwritearray str_buf access with get_str().
