+++ vcproj/map-server.vcxproj.filters
@@ -127,6 +127,9 @@
     <ClCompile Include="..\src\map\buildin_mob.c">
       <Filter>map_sql</Filter>
     </ClCompile>
+    <ClCompile Include="..\src\map\buildin_csv.c">
+      <Filter>map_sql</Filter>
+    </ClCompile>
     <ClCompile Include="..\src\map\skill.c">
       <Filter>map_sql</Filter>
     </ClCompile>
